+++ src/Compilers/CSharp/Portable/Binder/Binder_Statements.cs
@@ -20,7 +20,7 @@ namespace Microsoft.CodeAnalysis.CSharp
     internal partial class Binder
     {
         /// <summary>
-        /// This is the set of parameters and local variables that were used as arguments to 
+        /// This is the set of parameters and local variables that were used as arguments to
         /// lock or using statements in enclosing scopes.
         /// </summary>
         /// <remarks>
@@ -153,8 +153,8 @@ namespace Microsoft.CodeAnalysis.CSharp
 
             Debug.Assert(result.Syntax is StatementSyntax, "BoundStatement should be associated with a statement syntax.");
 
-            Debug.Assert(System.Linq.Enumerable.Contains(result.Syntax.AncestorsAndSelf(), node), @"Bound statement (or one of its parents) 
-                                                                            should have same syntax as the given syntax node. 
+            Debug.Assert(System.Linq.Enumerable.Contains(result.Syntax.AncestorsAndSelf(), node), @"Bound statement (or one of its parents)
+                                                                            should have same syntax as the given syntax node.
                                                                             Otherwise it may be confusing to the binder cache that uses syntax node as keys.");
 
             return result;
@@ -467,7 +467,7 @@ namespace Microsoft.CodeAnalysis.CSharp
             HashSet<DiagnosticInfo> useSiteDiagnostics = null;
             var binder = this.LookupSymbolsWithFallback(result, node.Identifier.ValueText, arity: 0, useSiteDiagnostics: ref useSiteDiagnostics, options: LookupOptions.LabelsOnly);
 
-            // result.Symbols can be empty in some malformed code, e.g. when a labeled statement is used an embedded statement in an if or foreach statement    
+            // result.Symbols can be empty in some malformed code, e.g. when a labeled statement is used an embedded statement in an if or foreach statement
             // In this case we create new label symbol on the fly, and an error is reported by parser
             var symbol = result.Symbols.Count > 0 && result.IsMultiViable ?
                 (LabelSymbol)result.Symbols.First() :
@@ -776,7 +776,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
             if (isVar)
             {
-                // There are a number of ways in which a var decl can be illegal, but in these 
+                // There are a number of ways in which a var decl can be illegal, but in these
                 // cases we should report an error and then keep right on going with the inference.
 
                 if (isConst)
@@ -793,11 +793,11 @@ namespace Microsoft.CodeAnalysis.CSharp
                 // users showed that there was no consensus on whether the above should mean
                 // "double x = 10, y = 123.4;", taking the best type available and substituting
                 // that for "var", or treating it as "var x = 10; var y = 123.4;" -- since there
-                // was no consensus we decided to simply make it illegal. 
+                // was no consensus we decided to simply make it illegal.
                 //
                 // In dev10 for error recovery in the IDE we do an odd thing -- we simply take
                 // the type of the first variable and use it. So that is "int x = 10, y = 123.4;".
-                // 
+                //
                 // This seems less than ideal. In the error recovery scenario it probably makes
                 // more sense to treat that as "var x = 10; var y = 123.4;" and do each inference
                 // separately.
@@ -813,7 +813,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                 // In the native compiler when given a situation like
                 //
                 // D[] x;
-                // 
+                //
                 // where D is a static type we report both that D cannot be an element type
                 // of an array, and that D[] is not a valid type for a local variable.
                 // This seems silly; the first error is entirely sufficient. We no longer
@@ -867,7 +867,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
             BoundExpression expression = BindToNaturalType(BindValue(initializer, diagnostics, valueKind), diagnostics);
 
-            // Certain expressions (null literals, method groups and anonymous functions) have no type of 
+            // Certain expressions (null literals, method groups and anonymous functions) have no type of
             // their own and therefore cannot be the initializer of an implicitly typed local.
             if (!expression.HasAnyErrors && !expression.HasExpressionType())
             {
@@ -990,7 +990,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
                 initializerOpt = BindInferredVariableInitializer(diagnostics, value, valueKind, localSymbol.RefKind, declarator);
 
-                // If we got a good result then swap the inferred type for the "var" 
+                // If we got a good result then swap the inferred type for the "var"
                 TypeSymbol initializerType = initializerOpt?.Type;
                 if ((object)initializerType != null)
                 {
@@ -1030,7 +1030,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                     initializerOpt = BindPossibleArrayInitializer(value, declTypeOpt.Type, valueKind, diagnostics);
                     if (kind != LocalDeclarationKind.FixedVariable)
                     {
-                        // If this is for a fixed statement, we'll do our own conversion since there are some special cases.		
+                        // If this is for a fixed statement, we'll do our own conversion since there are some special cases.
                         initializerOpt = GenerateConversionForAssignment(
                             declTypeOpt.Type,
                             initializerOpt,
@@ -1091,9 +1091,9 @@ namespace Microsoft.CodeAnalysis.CSharp
 
             if (kind == LocalDeclarationKind.FixedVariable || kind == LocalDeclarationKind.UsingVariable)
             {
-                // CONSIDER: The error message is "you must provide an initializer in a fixed 
-                // CONSIDER: or using declaration". The error message could be targeted to 
-                // CONSIDER: the actual situation. "you must provide an initializer in a 
+                // CONSIDER: The error message is "you must provide an initializer in a fixed
+                // CONSIDER: or using declaration". The error message could be targeted to
+                // CONSIDER: the actual situation. "you must provide an initializer in a
                 // CONSIDER: 'fixed' declaration."
 
                 if (initializerOpt == null)
@@ -1167,8 +1167,8 @@ namespace Microsoft.CodeAnalysis.CSharp
 
         internal ImmutableArray<BoundExpression> BindDeclaratorArguments(VariableDeclaratorSyntax declarator, DiagnosticBag diagnostics)
         {
-            // It is possible that we have a bracketed argument list, like "int x[];" or "int x[123];" 
-            // in a non-fixed-size-array declaration . This is a common error made by C++ programmers. 
+            // It is possible that we have a bracketed argument list, like "int x[];" or "int x[123];"
+            // in a non-fixed-size-array declaration . This is a common error made by C++ programmers.
             // We have already given a good error at parse time telling the user to either make it "fixed"
             // or to move the brackets to the type. However, we should still do semantic analysis of
             // the arguments, so that errors in them are discovered, hovering over them in the IDE
@@ -1279,7 +1279,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                         break;
                     }
 
-                    // if the feature was enabled, but something went wrong with the method, report that, otherwise don't. 
+                    // if the feature was enabled, but something went wrong with the method, report that, otherwise don't.
                     // If feature is not enabled, additional errors would be just noise.
                     bool extensibleFixedEnabled = ((CSharpParseOptions)initializerOpt.SyntaxTree.Options)?.IsFeatureEnabled(MessageID.IDS_FeatureExtensibleFixedStatement) != false;
                     if (extensibleFixedEnabled)
@@ -1466,11 +1466,11 @@ namespace Microsoft.CodeAnalysis.CSharp
 
             if (!op1.HasAnyErrors)
             {
-                // Build bound conversion. The node might not be used if this is a dynamic conversion 
+                // Build bound conversion. The node might not be used if this is a dynamic conversion
                 // but diagnostics should be reported anyways.
                 var conversion = GenerateConversionForAssignment(op1.Type, op2, diagnostics, isRefAssignment: isRef);
 
-                // If the result is a dynamic assignment operation (SetMember or SetIndex), 
+                // If the result is a dynamic assignment operation (SetMember or SetIndex),
                 // don't generate the boxing conversion to the dynamic type.
                 // Leave the values as they are, and deal with the conversions at runtime.
                 if (op1.Kind != BoundKind.DynamicIndexerAccess &&
@@ -1621,7 +1621,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                 propertySymbol = propertySymbol.OriginalDefinition;
             }
 
-            var sourceProperty = propertySymbol as SourceOrRecordPropertySymbol;
+            var sourceProperty = propertySymbol as SourcePropertySymbolBase;
             var propertyIsStatic = propertySymbol.IsStatic;
 
             return (object)sourceProperty != null &&
@@ -1766,13 +1766,13 @@ namespace Microsoft.CodeAnalysis.CSharp
             Debug.Assert((object)targetType != null);
             Debug.Assert(expression != null);
 
-            // We wish to avoid "cascading" errors, so if the expression we are 
+            // We wish to avoid "cascading" errors, so if the expression we are
             // attempting to convert to a type had errors, suppress additional
-            // diagnostics. However, if the expression 
+            // diagnostics. However, if the expression
             // with errors is an unbound lambda then the errors are almost certainly
             // syntax errors. For error recovery analysis purposes we wish to bind
             // error lambdas like "Action<int> f = x=>{ x. };" because IntelliSense
-            // needs to know that x is of type int. 
+            // needs to know that x is of type int.
 
             if (expression.HasAnyErrors && expression.Kind != BoundKind.UnboundLambda)
             {
@@ -1796,7 +1796,7 @@ namespace Microsoft.CodeAnalysis.CSharp
             }
             else if (!conversion.IsImplicit || !conversion.IsValid)
             {
-                // We suppress conversion errors on default parameters; eg, 
+                // We suppress conversion errors on default parameters; eg,
                 // if someone says "void M(string s = 123) {}". We will report
                 // a special error in the default parameter binder.
 
@@ -1832,7 +1832,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
             var reason = Conversions.IsAnonymousFunctionCompatibleWithType(anonymousFunction, targetType);
 
-            // It is possible that the conversion from lambda to delegate is just fine, and 
+            // It is possible that the conversion from lambda to delegate is just fine, and
             // that we ended up here because the target type, though itself is not an error
             // type, contains a type argument which is an error type. For example, converting
             // (Goo goo)=>{} to Action<Goo> is a perfectly legal conversion even if Goo is undefined!
@@ -1875,7 +1875,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
             var delegateType = targetType.GetDelegateType();
 
-            // The target type is a valid delegate or expression tree type. Is there something wrong with the 
+            // The target type is a valid delegate or expression tree type. Is there something wrong with the
             // parameter list?
 
             // First off, is there a parameter list at all?
@@ -1889,7 +1889,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                 // D d = delegate {};
                 //
                 // error CS1676: Parameter 1 must be declared with the 'out' keyword
-                // error CS1688: Cannot convert anonymous method block without a parameter list 
+                // error CS1688: Cannot convert anonymous method block without a parameter list
                 // to delegate type 'D' because it has one or more out parameters
                 //
                 // This seems redundant, (because there is no "parameter 1" in the source code)
@@ -1910,7 +1910,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
             // The parameter list exists and had the right number of parameters. Were any of its types bad?
 
-            // If any parameter type of the lambda is an error type then suppress 
+            // If any parameter type of the lambda is an error type then suppress
             // further errors. We've already reported errors on the bad type.
             if (anonymousFunction.HasExplicitlyTypedParameterList)
             {
@@ -2041,7 +2041,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                         // CLEVERNESS: By passing ConstantValue.Bad, we tell HasImplicitConstantExpressionConversion to ignore the constant
                         // value and only consider the types.
 
-                        // If there would be an implicit constant conversion for a different constant of the same type 
+                        // If there would be an implicit constant conversion for a different constant of the same type
                         // (i.e. one that's not out of range), then it's more helpful to report the range check failure
                         // than to suggest inserting a cast.
                         Error(diagnostics, ErrorCode.ERR_ConstOutOfRange, syntax, sourceConstantValueOpt.Value, targetType);
@@ -2266,9 +2266,9 @@ namespace Microsoft.CodeAnalysis.CSharp
 
             // report all leaf elements of the tuple literal that failed to convert
             // NOTE: we are not responsible for reporting use site errors here, just the failed leaf conversions.
-            // By the time we get here we have done analysis and know we have failed the cast in general, and diagnostics collected in the process is already in the bag. 
+            // By the time we get here we have done analysis and know we have failed the cast in general, and diagnostics collected in the process is already in the bag.
             // The only thing left is to form a diagnostics about the actually failing conversion(s).
-            // This whole method does not itself collect any usesite diagnostics. Its only purpose is to produce an error better than "conversion failed here"           
+            // This whole method does not itself collect any usesite diagnostics. Its only purpose is to produce an error better than "conversion failed here"
             HashSet<DiagnosticInfo> usDiagsUnused = null;
 
             for (int i = 0; i < targetElementTypes.Length; i++)
@@ -2296,24 +2296,24 @@ namespace Microsoft.CodeAnalysis.CSharp
 
         internal BoundExpression BindBooleanExpression(ExpressionSyntax node, DiagnosticBag diagnostics)
         {
-            // SPEC: 
-            // A boolean-expression is an expression that yields a result of type bool; 
-            // either directly or through application of operator true in certain 
+            // SPEC:
+            // A boolean-expression is an expression that yields a result of type bool;
+            // either directly or through application of operator true in certain
             // contexts as specified in the following.
             //
-            // The controlling conditional expression of an if-statement, while-statement, 
-            // do-statement, or for-statement is a boolean-expression. The controlling 
-            // conditional expression of the ?: operator follows the same rules as a 
+            // The controlling conditional expression of an if-statement, while-statement,
+            // do-statement, or for-statement is a boolean-expression. The controlling
+            // conditional expression of the ?: operator follows the same rules as a
             // boolean-expression, but for reasons of operator precedence is classified
             // as a conditional-or-expression.
             //
-            // A boolean-expression is required to be implicitly convertible to bool 
-            // or of a type that implements operator true. If neither requirement 
+            // A boolean-expression is required to be implicitly convertible to bool
+            // or of a type that implements operator true. If neither requirement
             // is satisfied, a binding-time error occurs.
             //
-            // When a boolean expression cannot be implicitly converted to bool but does 
-            // implement operator true, then following evaluation of the expression, 
-            // the operator true implementation provided by that type is invoked 
+            // When a boolean expression cannot be implicitly converted to bool but does
+            // implement operator true, then following evaluation of the expression,
+            // the operator true implementation provided by that type is invoked
             // to produce a bool value.
             //
             // SPEC ERROR: The third paragraph above is obviously not correct; we need
@@ -2420,7 +2420,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                 diagnostics: diagnostics);
 
             // Consider op_true to be compiler-generated so that it doesn't appear in the semantic model.
-            // UNDONE: If we decide to expose the operator in the semantic model, we'll have to remove the 
+            // UNDONE: If we decide to expose the operator in the semantic model, we'll have to remove the
             // WasCompilerGenerated flag (and possibly suppress the symbol in specific APIs).
             return new BoundUnaryOperator(node, signature.Kind, resultOperand, ConstantValue.NotAvailable, signature.Method, resultKind, originalUserDefinedOperators, signature.ReturnType)
             {
@@ -2511,7 +2511,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
             if (isVar && count > 1)
             {
-                // There are a number of ways in which a var decl can be illegal, but in these 
+                // There are a number of ways in which a var decl can be illegal, but in these
                 // cases we should report an error and then keep right on going with the inference.
 
                 Error(diagnostics, ErrorCode.ERR_ImplicitlyTypedVariableMultipleDeclarator, nodeOpt);
@@ -2729,8 +2729,8 @@ namespace Microsoft.CodeAnalysis.CSharp
                 return new BoundReturnStatement(syntax, refKind, BindToTypeForErrorRecovery(arg), hasErrors: true);
             }
 
-            // The return type could be null; we might be attempting to infer the return type either 
-            // because of method type inference, or because we are attempting to do error analysis 
+            // The return type could be null; we might be attempting to infer the return type either
+            // because of method type inference, or because we are attempting to do error analysis
             // on a lambda expression of unknown return type.
             if ((object)retType != null)
             {
@@ -2742,7 +2742,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                         var lambda = container as LambdaSymbol;
                         if ((object)lambda != null)
                         {
-                            // Error case: void-returning or async task-returning method or lambda with "return x;" 
+                            // Error case: void-returning or async task-returning method or lambda with "return x;"
                             var errorCode = retType.IsVoidType()
                                 ? ErrorCode.ERR_RetNoObjectRequiredLambda
                                 : ErrorCode.ERR_TaskRetNoObjectRequiredLambda;
@@ -2754,12 +2754,12 @@ namespace Microsoft.CodeAnalysis.CSharp
                             // COMPATIBILITY: The native compiler also produced an error
                             // COMPATIBILITY: "Cannot convert lambda expression to delegate type 'Action' because some of the
                             // COMPATIBILITY: return types in the block are not implicitly convertible to the delegate return type"
-                            // COMPATIBILITY: This error doesn't make sense in the "void" case because the whole idea of 
+                            // COMPATIBILITY: This error doesn't make sense in the "void" case because the whole idea of
                             // COMPATIBILITY: "conversion to void" is a bit unusual, and we've already given a good error.
                         }
                         else
                         {
-                            // Error case: void-returning or async task-returning method or lambda with "return x;" 
+                            // Error case: void-returning or async task-returning method or lambda with "return x;"
                             var errorCode = retType.IsVoidType()
                                 ? ErrorCode.ERR_RetNoObjectRequired
                                 : ErrorCode.ERR_TaskRetNoObjectRequired;
@@ -2963,7 +2963,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
                     // If the previous type is a generic parameter we don't know what exception types it's gonna catch exactly.
                     // If it is a class-type we know it's gonna catch all exception types of its type and types that are derived from it.
-                    // So if the current type is a class-type (or an effective base type of a generic parameter) 
+                    // So if the current type is a class-type (or an effective base type of a generic parameter)
                     // that derives from the previous type the current catch is unreachable.
 
                     if (previousBlock.ExceptionFilterOpt == null && (object)previousType != null && !previousType.IsErrorType())
@@ -3111,7 +3111,7 @@ namespace Microsoft.CodeAnalysis.CSharp
         }
 
         /// <summary>
-        /// Wrap a given expression e into a block as either { e; } or { return e; } 
+        /// Wrap a given expression e into a block as either { e; } or { return e; }
         /// Shared between lambda and expression-bodied method binding.
         /// </summary>
         internal BoundBlock CreateBlockFromExpression(CSharpSyntaxNode node, ImmutableArray<LocalSymbol> locals, RefKind refKind, BoundExpression expression, ExpressionSyntax expressionSyntax, DiagnosticBag diagnostics)
@@ -3422,7 +3422,7 @@ namespace Microsoft.CodeAnalysis.CSharp
         }
 
         /// <summary>
-        /// Perform a lookup for the specified method on the specified expression by attempting to invoke it 
+        /// Perform a lookup for the specified method on the specified expression by attempting to invoke it
         /// </summary>
         /// <param name="receiver">The expression to perform pattern lookup on</param>
         /// <param name="methodName">Method to search for.</param>
