+++ src/Compilers/CSharp/Portable/Compiler/MethodBodySynthesizer.cs
@@ -12,7 +12,7 @@ using Microsoft.CodeAnalysis.RuntimeMembers;
 namespace Microsoft.CodeAnalysis.CSharp
 {
     /// <summary>
-    /// Contains methods related to synthesizing bound nodes in initial binding 
+    /// Contains methods related to synthesizing bound nodes in initial binding
     /// form that needs lowering, primarily method bodies for compiler-generated methods.
     /// </summary>
     internal static class MethodBodySynthesizer
@@ -25,8 +25,8 @@ namespace Microsoft.CodeAnalysis.CSharp
         {
             // Script field initializers have to be emitted after the call to the base constructor because they can refer to "this" instance.
             //
-            // Unlike regular field initializers, initializers of global script variables can access "this" instance. 
-            // If the base class had a constructor that initializes its state a global variable would access partially initialized object. 
+            // Unlike regular field initializers, initializers of global script variables can access "this" instance.
+            // If the base class had a constructor that initializes its state a global variable would access partially initialized object.
             // For this reason Script class must always derive directly from a class that has no state (System.Object).
 
             SyntaxNode syntax = loweredBody.Syntax;
@@ -172,7 +172,7 @@ namespace Microsoft.CodeAnalysis.CSharp
         {
             Debug.Assert(accessor.MethodKind == MethodKind.PropertyGet || accessor.MethodKind == MethodKind.PropertySet);
 
-            var property = (SourcePropertySymbol)accessor.AssociatedSymbol;
+            var property = (SourcePropertySymbolBase)accessor.AssociatedSymbol;
             CSharpSyntaxNode syntax = property.CSharpSyntaxNode;
             BoundExpression thisReference = null;
             if (!accessor.IsStatic)
@@ -219,10 +219,10 @@ namespace Microsoft.CodeAnalysis.CSharp
 
         /// <summary>
         /// Generate a thread-safe accessor for a WinRT field-like event.
-        /// 
+        ///
         /// Add:
         ///   return EventRegistrationTokenTable&lt;Event&gt;.GetOrCreateEventRegistrationTokenTable(ref _tokenTable).AddEventHandler(value);
-        /// 
+        ///
         /// Remove:
         ///   EventRegistrationTokenTable&lt;Event&gt;.GetOrCreateEventRegistrationTokenTable(ref _tokenTable).RemoveEventHandler(value);
         /// </summary>
@@ -302,7 +302,7 @@ namespace Microsoft.CodeAnalysis.CSharp
             {
                 // {
                 //     return EventRegistrationTokenTable<Event>.GetOrCreateEventRegistrationTokenTable(ref _tokenTable).AddHandler(value);
-                // }   
+                // }
                 BoundStatement returnStatement = BoundReturnStatement.Synthesized(syntax, RefKind.None, processHandlerCall);
                 return BoundBlock.SynthesizedNoLocals(syntax, returnStatement);
             }
@@ -311,7 +311,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                 // {
                 //     EventRegistrationTokenTable<Event>.GetOrCreateEventRegistrationTokenTable(ref _tokenTable).RemoveHandler(value);
                 //     return;
-                // }  
+                // }
                 BoundStatement callStatement = new BoundExpressionStatement(syntax, processHandlerCall);
                 BoundStatement returnStatement = new BoundReturnStatement(syntax, RefKind.None, expressionOpt: null);
                 return BoundBlock.SynthesizedNoLocals(syntax, callStatement, returnStatement);
@@ -320,7 +320,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
         /// <summary>
         /// Generate a thread-safe accessor for a regular field-like event.
-        /// 
+        ///
         /// DelegateType tmp0 = _event; //backing field
         /// DelegateType tmp1;
         /// DelegateType tmp2;
@@ -329,12 +329,12 @@ namespace Microsoft.CodeAnalysis.CSharp
         ///     tmp2 = (DelegateType)Delegate.Combine(tmp1, value); //Remove for -=
         ///     tmp0 = Interlocked.CompareExchange&lt;DelegateType&gt;(ref _event, tmp2, tmp1);
         /// } while ((object)tmp0 != (object)tmp1);
-        /// 
+        ///
         /// Note, if System.Threading.Interlocked.CompareExchange&lt;T&gt; is not available,
         /// we emit the following code and mark the method Synchronized (unless it is a struct).
-        /// 
+        ///
         /// _event = (DelegateType)Delegate.Combine(_event, value); //Remove for -=
-        /// 
+        ///
         /// </summary>
         internal static BoundBlock ConstructFieldLikeEventAccessorBody_Regular(SourceEventSymbol eventSymbol, bool isAddMethod, CSharpCompilation compilation, DiagnosticBag diagnostics)
         {
@@ -522,7 +522,7 @@ namespace Microsoft.CodeAnalysis.CSharp
             Debug.Assert(method.MethodKind == MethodKind.Destructor);
             Debug.Assert(syntax.Kind() == SyntaxKind.Block || syntax.Kind() == SyntaxKind.ArrowExpressionClause);
 
-            // If this is a destructor and a base type has a Finalize method (see GetBaseTypeFinalizeMethod for exact 
+            // If this is a destructor and a base type has a Finalize method (see GetBaseTypeFinalizeMethod for exact
             // requirements), then we need to call that method in a finally block.  Otherwise, just return block as-is.
             // NOTE: the Finalize method need not be a destructor or be overridden by the current method.
             MethodSymbol baseTypeFinalize = GetBaseTypeFinalizeMethod(method);
@@ -572,11 +572,11 @@ namespace Microsoft.CodeAnalysis.CSharp
         }
 
         /// <summary>
-        /// Look for a base type method named "Finalize" that is protected (or protected internal), has no parameters, 
+        /// Look for a base type method named "Finalize" that is protected (or protected internal), has no parameters,
         /// and returns void.  It doesn't need to be virtual or a destructor.
         /// </summary>
         /// <remarks>
-        /// You may assume that this would share code and logic with PEMethodSymbol.OverridesRuntimeFinalizer, 
+        /// You may assume that this would share code and logic with PEMethodSymbol.OverridesRuntimeFinalizer,
         /// but FUNCBRECCS::bindDestructor has its own loop that performs these checks (differently).
         /// </remarks>
         private static MethodSymbol GetBaseTypeFinalizeMethod(MethodSymbol method)
